Use mergebounce to allow writes to the client-side store to be batched
The `Storage` constructor now takes an additional boolean parameter
`batchedWrites`.

[mergebounce](https://github.com/conversejs/mergebounce) is used to batch writes together.

`items` sent to the localForage `setItems` method is now an Object instead
of an Array so that multiple writes for the same model can be properly
merged. With an Array we'd either lose previous entries (default behavior)
or we'd only be able to append new data without removing data that needs
to be removed (when `contactArrays` is set to true).